requirements.txt
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
click==8.0.1
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Mail==0.9.1
flask-talisman==0.8.1
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
pycparser==2.20
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.1
